project_manager_mcp/requirements.txt
fastapi
uvicorn
pymongo
psycopg2-binary
python-dotenv
langchain-core
